refactor: move middlewares para pasta certa
