+++ config/expressConfig.ts
@@ -5,7 +5,7 @@ import cookieParser from 'cookie-parser';
 import routerUsuario from '../src/domains/Usuario/controllers/controllerUsuario.js';
 import routerArtista from '../src/domains/Artista/controllers/controllerArtista.js';
 import routerMusica from '../src/domains/Musica/controllers/controllerMusica.js';
-import { errorHandler } from '../src/middleware/errorHandler.js';
+import { errorHandler } from '../src/middlewares/errorHandler.js';
 
 dotenv.config();
