+++ src/domains/Artista/controllers/controllerArtista.ts
@@ -2,7 +2,7 @@ import { Router, Request, Response, NextFunction } from 'express';
 import { Prisma } from '@prisma/client';
 import serviceArtista from '../services/serviceArtista.js';
 import statusCodes from '../../../../utils/constants/statusCodes.js';
-import { validateId } from '../../../middleware/validateId.js';
+import { validateId } from '../../../middlewares/validateId.js';
 
 const router = Router();
